Makefile: don't remove newline at the end of templates
This was done for aesthetics reasons, but it changes the templates every
time since most text editors add a newline at the end.

Also, with the new commit hook introduced in 882419ab6, it leaves a
dirty git worktree after the commit.

Just remove it since we don't care that much about an extra line in the
resulting cloudconfig.