Moves extension units to top of cloudconfig (#60)
Towards https://github.com/giantswarm/aws-operator/issues/131

coreos-cloudinit starts unit sequentially, so the order of the
units needs to avoid deadlocks. In the case where the
aws-operator adds decrypt-tls-assets which is a requirement of
the later units, this causes a deadlock.

This changset moves the extension units to the top of the
cloudconfig, to avoid these deadlocks.