Add patch release v3.6.1 with fix to k8s apiserver
Kubernetes apiserver has a memory leak in v1.11.1. This can be mitigated
by applying a fix from apimachinery, but this is not merged in upstream
v1.11.x release yet. There is a custom built GiantSwarm release for this
(as AMD64 hyperkube image) that can be used while waiting for new
upstream release.